Get bazel tests working on Spark 3.2.
