+++ .bazelversion
@@ -1,0 +1,1 @@
+6.4.0
