+++ .ijwb/.bazelproject
@@ -1,16 +1,1 @@
-directories:
-  # Add the directories you want added as source here
-  # By default, we've added your entire workspace ('.')
-  .
-
-# Automatically includes all relevant targets under the 'directories' above
-derive_targets_from_directories: true
-
-targets:
-  # If source code isn't resolving, add additional targets that compile it here
-
-additional_languages:
-  # Uncomment any additional languages you want supported
-  python
-  scala
-  java
+import tools/ide_support/intellij/default_view.bazelproject
