+++ .ijwb/aspects/BUILD.bazel
@@ -1,0 +1,16 @@
+#
+# Description:
+# The final form of the BUILD file accessed at runtime as an external WORKSPACE.
+#
+
+licenses(["notice"])  # Apache 2.0
+
+load(":flag_hack.bzl", "define_flag_hack")
+
+exports_files([
+    "tools/PackageParser_deploy.jar",
+    "tools/CreateAar_deploy.jar",
+    "tools/JarFilter_deploy.jar",
+])
+
+define_flag_hack()
