+++ aggregator/src/test/scala/ai/chronon/aggregator/test/TwoStackLiteAggregatorTest.scala
@@ -22,12 +22,14 @@ import ai.chronon.aggregator.windowing.{TwoStackLiteAggregator, TwoStackLiteAggr
 import ai.chronon.api.{Aggregation, Builders, IntType, LongType, Operation, StructField, StructType, TimeUnit, Window}
 import junit.framework.TestCase
 import org.junit.Assert._
+import org.junit.Test
 import ai.chronon.api.Extensions.AggregationOps
 import com.google.gson.Gson
 
 import scala.collection.Seq
 
 class TwoStackLiteAggregatorTest extends TestCase{
+  @Test
   def testBufferWithTopK(): Unit = {
     val topK = new TopK[Integer](IntType, 2)
     val bankersBuffer = new TwoStackLiteAggregationBuffer(topK, 5)
@@ -53,6 +55,7 @@ class TwoStackLiteAggregatorTest extends TestCase{
     assertBufferEquals(Seq(10), bankersBuffer.query)
   }
 
+  @Test
   def testAgainstSawtooth(): Unit = {
     val timer = new Timer
     val queries = CStream.genTimestamps(new Window(30, TimeUnit.DAYS), 100000, 5 * 60 * 1000)
