+++ spark/BUILD.bazel
@@ -57,6 +57,7 @@ scala_library(
         "//api:api-models",
         "//online",
         "//third_party/java/spark:spark-libs",
+        "@bazel_tools//tools/java/runfiles",
         maven_artifact("com.google.code.gson:gson"),
         maven_artifact("org.apache.thrift:libthrift"),
         maven_artifact("com.google.guava:guava"),
@@ -79,19 +80,30 @@ scala_library(
     ],
 )
 
-scala_test_suite(
+scala_junit_test_suite(
     name = "test",
+    timeout = "eternal",
     srcs = glob(["src/test/scala/ai/chronon/spark/test/**/*.scala"]),
-    visibility = ["//visibility:public"],
+    data = [
+        "//spark/src/test/resources",
+        "//tools/policies:derby.policy",
+    ],
+    jvm_flags = [
+        "-Djava.security.policy=$(location //tools/policies:derby.policy)",
+    ],
+    resources = ["//spark/src/test/resources"],
+    strip_prefix = "src/test/scala/ai/chronon/spark/test",
+    visibility = ["//visibility:private"],
     deps = [
         ":spark",
-        ":test-lib",
         "//aggregator",
         "//aggregator:test-lib",
         "//api:api-lib",
         "//api:api-models",
         "//online",
         "//third_party/java/spark:spark-libs",
+        "//third_party/java/spark:spark-test-libs",
+        "@bazel_tools//tools/java/runfiles",
         scala_artifact("org.scala-lang.modules:scala-java8-compat"),
         maven_artifact("junit:junit"),
         maven_artifact("com.novocode:junit-interface"),
@@ -100,6 +112,11 @@ scala_test_suite(
         scala_artifact("com.fasterxml.jackson.module:jackson-module-scala"),
         maven_artifact("com.google.code.gson:gson"),
         scala_artifact("org.rogach:scallop"),
+        scala_artifact("org.scalactic:scalactic"),
+        scala_artifact("org.scalatest:scalatest-matchers-core"),
+        scala_artifact("org.scalatest:scalatest-core"),
+        maven_artifact("org.scalatest:scalatest-compatible"),
+        scala_artifact("org.scalatest:scalatest-shouldmatchers"),
         scala_artifact("org.scalatestplus:mockito-3-4"),
         maven_artifact("org.mockito:mockito-core"),
         maven_artifact("org.slf4j:slf4j-api"),
