+++ spark/src/main/scala/ai/chronon/spark/SparkSessionBuilder.scala
@@ -58,7 +58,12 @@ object SparkSessionBuilder {
 
     if (local) {
       //required to run spark locally with hive support enabled - for sbt test
-      System.setSecurityManager(null)
+      try {
+        System.setSecurityManager(null)
+      } catch {
+        case (t: java.lang.SecurityException) if t.getMessage.contains("GoogleTestSecurityManager") =>
+          // Running on Bazel, allow it.
+      }
     }
     val userName = Properties.userName
     val warehouseDir = localWarehouseLocation.map(expandUser).getOrElse(DefaultWarehouseDir.getAbsolutePath)
@@ -102,6 +107,7 @@ object SparkSessionBuilder {
         .config("spark.sql.warehouse.dir", s"$warehouseDir/data")
         .config("spark.hadoop.javax.jdo.option.ConnectionURL", metastoreDb)
         .config("spark.driver.bindAddress", "127.0.0.1")
+        .config("spark.ui.enabled", "false")
     } else {
       // hive jars need to be available on classpath - no needed for local testing
       baseBuilder
@@ -131,6 +137,7 @@ object SparkSessionBuilder {
         .master("local[*]")
         .config("spark.local.dir", s"/tmp/$userName/chronon-spark-streaming")
         .config("spark.kryo.registrationRequired", "true")
+        .config("spark.ui.enabled", "false")
     } else {
       baseBuilder
     }
