+++ spark/src/test/resources/BUILD.bazel
@@ -1,0 +1,5 @@
+filegroup(
+    name = "resources",
+    srcs = glob(["**"]),
+    visibility = ["//visibility:public"],
+)
