+++ spark/src/test/scala/ai/chronon/spark/test/FetcherTest.scala
@@ -33,6 +33,7 @@ import org.apache.spark.sql.catalyst.expressions.GenericRow
 import org.apache.spark.sql.functions.{avg, col, lit}
 import org.apache.spark.sql.{DataFrame, Row, SparkSession}
 import org.junit.Assert.{assertEquals, assertTrue}
+import org.junit.Test
 import org.mockito.ArgumentMatchers.{any, anyString}
 import org.mockito.Mockito.{reset, spy, when}
 import org.slf4j.LoggerFactory
@@ -655,12 +656,14 @@ class FetcherTest extends TestCase {
     assertEquals(0, diff.count())
   }
 
+  @Test
   def testTemporalFetchJoinDeterministic(): Unit = {
     val namespace = "deterministic_fetch"
     val joinConf = generateMutationData(namespace)
     compareTemporalFetch(joinConf, "2021-04-10", namespace, consistencyCheck = false, dropDsOnWrite = true)
   }
 
+  @Test
   def testTemporalFetchJoinDerivation(): Unit = {
     val namespace = "derivation_fetch"
     val joinConf = generateMutationData(namespace)
@@ -675,6 +678,7 @@ class FetcherTest extends TestCase {
     compareTemporalFetch(joinConf, "2021-04-10", namespace, consistencyCheck = false, dropDsOnWrite = true)
   }
 
+  @Test
   def testTemporalFetchJoinDerivationRenameOnly(): Unit = {
     val namespace = "derivation_fetch_rename_only"
     val joinConf = generateMutationData(namespace)
@@ -685,6 +689,7 @@ class FetcherTest extends TestCase {
     compareTemporalFetch(joinConf, "2021-04-10", namespace, consistencyCheck = false, dropDsOnWrite = true)
   }
 
+  @Test
   def testTemporalFetchJoinGenerated(): Unit = {
     val namespace = "generated_fetch"
     val joinConf = generateRandomData(namespace)
@@ -695,6 +700,7 @@ class FetcherTest extends TestCase {
                          dropDsOnWrite = false)
   }
 
+  @Test
   def testTemporalTiledFetchJoinDeterministic(): Unit = {
     val namespace = "deterministic_tiled_fetch"
     val joinConf = generateEventOnlyData(namespace, groupByCustomJson = Some("{\"enable_tiling\": true}"))
@@ -702,6 +708,7 @@ class FetcherTest extends TestCase {
   }
 
   // test soft-fail on missing keys
+  @Test
   def testEmptyRequest(): Unit = {
     val spark: SparkSession = createSparkSession()
     val namespace = "empty_request"
@@ -727,6 +734,7 @@ class FetcherTest extends TestCase {
     assertTrue(responseMap.keys.forall(_.endsWith("_exception")))
   }
 
+  @Test
   def testTemporalFetchGroupByNonExistKey(): Unit = {
     val namespace = "non_exist_key_group_by_fetch"
     val spark: SparkSession = createSparkSession()
@@ -752,6 +760,7 @@ class FetcherTest extends TestCase {
     assertEquals(expected, result.head.values.get)
   }
 
+  @Test
   def testKVStorePartialFailure(): Unit = {
 
     val spark: SparkSession = createSparkSession()
@@ -781,6 +790,7 @@ class FetcherTest extends TestCase {
     exceptionKeys.foreach(k => assertTrue(responseMap.contains(k)))
   }
 
+  @Test
   def testGroupByServingInfoTtlCacheRefresh(): Unit = {
     val namespace = "test_group_by_serving_info_ttl_cache_refresh"
     val spark: SparkSession = createSparkSession()
@@ -817,6 +827,7 @@ class FetcherTest extends TestCase {
     assertTrue(response3.values.isSuccess)
   }
 
+  @Test
   def testJoinConfTtlCacheRefresh(): Unit = {
     val namespace = "test_join_conf_ttl_cache_refresh"
     val spark: SparkSession = createSparkSession()
@@ -849,105 +860,5 @@ class FetcherTest extends TestCase {
     reset(spyKvStore)
     val response2 = fetch()
     assertTrue(response2.isSuccess)
-  }
-}
-
-object FetcherTestUtil {
-  @transient lazy val logger = LoggerFactory.getLogger(getClass)
-  def joinResponses(spark: SparkSession,
-                    requests: Array[Request],
-                    mockApi: MockApi,
-                    useJavaFetcher: Boolean = false,
-                    runCount: Int = 1,
-                    samplePercent: Double = -1,
-                    logToHive: Boolean = false,
-                    debug: Boolean = false)(implicit ec: ExecutionContext): (List[Response], DataFrame) = {
-    val chunkSize = 100
-    @transient lazy val fetcher = mockApi.buildFetcher(debug)
-    @transient lazy val javaFetcher = mockApi.buildJavaFetcher()
-
-    def fetchOnce = {
-      var latencySum: Long = 0
-      var latencyCount = 0
-      val blockStart = System.currentTimeMillis()
-      val result = requests.iterator
-        .grouped(chunkSize)
-        .map { oldReqs =>
-          // deliberately mis-type a few keys
-          val r = oldReqs
-            .map(r =>
-              r.copy(keys = r.keys.mapValues { v =>
-                if (v.isInstanceOf[java.lang.Long]) v.toString else v
-              }.toMap))
-          val responses = if (useJavaFetcher) {
-            // Converting to java request and using the toScalaRequest functionality to test conversion
-            val convertedJavaRequests = r.map(new JavaRequest(_)).toJava
-            val javaResponse = javaFetcher.fetchJoin(convertedJavaRequests)
-            FutureConverters
-              .toScala(javaResponse)
-              .map(
-                _.toScala.map(jres =>
-                  Response(
-                    Request(jres.request.name, jres.request.keys.toScala.toMap, Option(jres.request.atMillis)),
-                    jres.values.toScala.map(_.toScala)
-                  )))
-          } else {
-            fetcher.fetchJoin(r)
-          }
-
-          // fix mis-typed keys in the request
-          val fixedResponses =
-            responses.map(resps => resps.zip(oldReqs).map { case (resp, req) => resp.copy(request = req) })
-          System.currentTimeMillis() -> fixedResponses
-        }
-        .flatMap {
-          case (start, future) =>
-            val result = Await.result(future, Duration(10000, SECONDS)) // todo: change back to millis
-            val latency = System.currentTimeMillis() - start
-            latencySum += latency
-            latencyCount += 1
-            result
-        }
-        .toList
-      val latencyMillis = latencySum.toFloat / latencyCount.toFloat
-      val qps = (requests.length * 1000.0) / (System.currentTimeMillis() - blockStart).toFloat
-      (latencyMillis, qps, result)
-    }
-
-    // to overwhelm the profiler with fetching code path
-    // so as to make it prominent in the flamegraph & collect enough stats
-
-    var latencySum = 0.0
-    var qpsSum = 0.0
-    var loggedValues: Seq[LoggableResponseBase64] = null
-    var result: List[Response] = null
-    (0 until runCount).foreach { _ =>
-      val (latency, qps, resultVal) = fetchOnce
-      result = resultVal
-      loggedValues = mockApi.flushLoggedValues
-      latencySum += latency
-      qpsSum += qps
-    }
-    val fetcherNameString = if (useJavaFetcher) "Java" else "Scala"
-
-    logger.info(s"""
-         |Averaging fetching stats for $fetcherNameString Fetcher over ${requests.length} requests $runCount times
-         |with batch size: $chunkSize
-         |average qps: ${qpsSum / runCount}
-         |average latency: ${latencySum / runCount}
-         |""".stripMargin)
-    val loggedDf = mockApi.loggedValuesToDf(loggedValues, spark)
-    if (logToHive) {
-      TableUtils(spark).insertPartitions(
-        loggedDf,
-        mockApi.logTable,
-        partitionColumns = Seq("ds", "name")
-      )
-    }
-    if (samplePercent > 0) {
-      logger.info(s"logged count: ${loggedDf.count()}")
-      loggedDf.show()
-    }
-    result -> loggedDf
   }
 }
