+++ tools/build_rules/prelude_bazel
@@ -8,6 +8,7 @@ load(
     "scala_jar")
 
 load("@rules_java//java:defs.bzl", "java_library","java_binary")
-load("@io_bazel_rules_scala//scala:scala.bzl", "scala_library", "scala_binary","scala_test_suite")
+load("@io_bazel_rules_scala//scala:scala.bzl", "scala_library", "scala_binary", "scala_junit_test", "scala_test_suite")
 load("//tools/build_rules:maven_artifact.bzl", "maven_artifact", "scala_artifact")
 load("//tools/build_rules:jvm_binary.bzl", "jvm_binary")
+load("//tools/build_rules:testing.bzl", "scala_junit_test_suite")
